Introduce Charging Limiter
This is a Kernel based charging limiter. It's version 1 so just some dirty code
I will update it later to don't hurt anyones eyes.
Also the feature will be hidden for now since it has some issues

Issue behavior
Charged till (at example) 80%
-> Charging suspended
Discharging till 79%
-> Charging continued

And this is repeating...
So I will fix this later, and reveal it for some testers.

Till now the process is working, so i will bring it with update but it's not perfect yet